indexer-agent: Make sure to allocate to all subgraphs worth indexing
When a channel is settled (e.g. by the indexer service or by the indexer-agent), this
removes the allocation but it doesn't make the subgraph no longer worth indexing. To
ensure that the indexer keeps serving the subgraph, we need to reallocate. By
automatically allocating on all subgraphs that we aim to index, we ensure that.