indexer-agent: Fix incorrect token/allocation amounts
The default internal representation is always the raw wei-style format,
i.e. very large numbers like `100000000000000.0`. For displaying this
has to go through `formatGRT`. Every time a user-provided GRT value is
dealt with, it needs to be parsed into a wei-style number with
`parseGRT`.

Values from the indexing rules and the subgraph/contracts don't have to
be parsed again, they have to be loaded with `BigNumber.from`.