indexer-agent: Never fail reconciliation early
We want the allocations for all deployments to be reconciled, even if
some of them e.g. fail to allocate or fail to index. This is especially
important when indexer are close to having all of their stake allocated.
Prior to this fix, the agent could end up trying to allocate over and
over again and never get to the point where it would close old
allocations and thereby free up some stake for creating those new
allocations.

On the side, this also fixes #126 by claiming rewards even if the
network is paused.