indexer-agent: Update to new `proofOfIndexing` API (#186)
This adds passing the epoch start block number to the `proofOfIndexing`
query, which is being added to graph-node in
https://github.com/graphprotocol/graph-node/pull/2100.

This change is necessary so that we can reliably obtain a POI even if
the graph-node block store doesn't have the Ethereum block in question.
This fixes the main problem in #186. For another improvement (allowing
indexers to deal with legitimate 0x0 POIs by closing allocations with
manual intervention), we'll create a separate issue.