*: Crash-tolerant transfer resolution and withdrawals
This tracks the state of allocations and transfers in the db more
accurately.

In order to facilitate crash tolerance, transfers are now resolved using
a priority queue (where the priority is the time when a transfer is
ready to be withdrawn; i.e. 10 minutes after its allocation has closed).

This has two benefits:

a) Assuming gateways stop using transfers for closed allocations and the
indexer service flushes their receipts to the db within these 10
minutes, there is no risk of missing out on even a single payment.

b) After a restart, we can push transfers that are ready to resolve back
into the processing priority queue. Thanks to this, there is no risk of
a transfer not resolving.

We add allocation summaries in the database to keep track of how many
transfers were created, resolved or failed to resolve, and are currently
open, for every single allocation. Only once an allocation has closed
and there are no open (i.e. unresolved) transfers left, we withdraw the
amount of query paymets received for the allocation.

Since the state channel balance says nothing about how much GRT each
allocation has received in query fees, we keep track of how many fees
were resolved via transfers at the allocation level, and how many fees
were withdrawn so far. This allows to withdraw multiple times. Now that
I think of it, I forgot why that was important. Either way, it just
makes things a little more convenient.